feat: 댓글, 게시글 좋아요 기능 구현 (#54)
* ㅇㅇㅇㅇ

* 댓글, 게시글 좋아요 기능 구현

* refactor:like 컨트롤러 commentLike 부분 수정

* remove firebase key

* fix:search 기능 복구

* fix:main이랑 다른 부분 수정

* refactor:LikeService,dto,converter 등 liketype 받아오는 방식 변경

* fix:likeRequest를 받지않고 liked는 enum으로 변경 + comment,post에 Long likeCount 추가

* refactor: 게시글 좋아요 PATCH로 수정 및 관련 구조 개선

- 게시글 좋아요 기능을 POST에서 PATCH 으로 변경
- Comment, Like, Vote 를 Post 내부에 통합
- Like의 liked 필드를 Enum → Boolean으로 변경
- 댓글 좋아요(commentLike)는 댓글 ID만 받도록 로직 변경

* fix:newsVoteService 수정

* fix:LikeController RequestMapping 주소 변경

---------
