--- src/main/java/com/tave/alarmissue/comment/domain/Comment.java
@@ -1,48 +1,0 @@
-package com.tave.alarmissue.comment.domain;
-
-import com.tave.alarmissue.global.domain.BaseTimeEntity;
-import com.tave.alarmissue.post.domain.Post;
-import com.tave.alarmissue.user.domain.UserEntity;
-import com.tave.alarmissue.vote.domain.Vote;
-import com.tave.alarmissue.vote.domain.VoteType;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.*;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
-
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Builder
-public class Comment extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long commentId;
-
-    @Column(columnDefinition = "TEXT", nullable = false)
-    private String commentContent;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private UserEntity user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id", nullable = false)
-    private Post post;
-
-    @Enumerated(EnumType.STRING)
-    @JoinColumn(nullable = true)
-    private VoteType voteTypeSnapshot;
-
-    @Builder
-    public Comment(Long commentId, String commentContent, UserEntity user, Post post,VoteType voteTypeSnapshot) {
-        this.commentId = commentId;
-        this.commentContent = commentContent;
-        this.user = user;
-        this.post = post;
-        this.voteTypeSnapshot = voteTypeSnapshot;
-    }
-}
