--- src/main/java/com/tave/alarmissue/comment/dto/request/CommentCreateRequestDto.java
@@ -1,10 +1,0 @@
-package com.tave.alarmissue.comment.dto.request;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class CommentCreateRequestDto {
-    private String commentContent;
-}
