--- src/main/java/com/tave/alarmissue/comment/dto/response/CommentResponseDto.java
@@ -1,20 +1,0 @@
-package com.tave.alarmissue.comment.dto.response;
-
-import com.tave.alarmissue.vote.domain.VoteType;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@AllArgsConstructor
-@Builder
-public class CommentResponseDto {
-    private Long commentId;
-    private String commentContent;
-    private String nickname;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-    private VoteType voteType;
-}
