--- src/main/java/com/tave/alarmissue/comment/exception/CommentException.java
@@ -1,10 +1,0 @@
-package com.tave.alarmissue.comment.exception;
-
-import com.tave.alarmissue.global.exception.CustomException;
-import com.tave.alarmissue.global.exception.ErrorCode;
-
-public class CommentException extends CustomException {
-    public CommentException(ErrorCode errorCode) {super(errorCode);}
-
-    public CommentException(ErrorCode errorCode, String message) {super(errorCode, message);}
-}
