--- src/main/java/com/tave/alarmissue/comment/repository/CommentRepository.java
@@ -1,12 +1,0 @@
-package com.tave.alarmissue.comment.repository;
-
-import com.tave.alarmissue.comment.domain.Comment;
-import com.tave.alarmissue.post.domain.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-    void deleteAllByPost(Post post);
-
-}
