+++ src/main/java/com/tave/alarmissue/news/service/NewsVoteService.java
@@ -8,18 +8,19 @@ import com.tave.alarmissue.news.dto.response.NewsVoteCountResponse;
 import com.tave.alarmissue.news.dto.response.NewsVoteResponseDto;
 import com.tave.alarmissue.news.repository.NewsVoteRepository;
 import com.tave.alarmissue.news.repository.NewsRepository;
-import com.tave.alarmissue.post.domain.Post;
+import com.tave.alarmissue.news.exceptions.NewsVoteException;
+import com.tave.alarmissue.post.exception.VoteException;
 import com.tave.alarmissue.user.domain.UserEntity;
 import com.tave.alarmissue.user.repository.UserRepository;
-import com.tave.alarmissue.vote.exception.VoteException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-import static com.tave.alarmissue.vote.exception.VoteErrorCode.POST_ID_NOT_FOUND;
-import static com.tave.alarmissue.vote.exception.VoteErrorCode.USER_ID_NOT_FOUND;
+import static com.tave.alarmissue.news.exceptions.VoteErrorCode.NEWS_ID_NOT_FOUND;
+import static com.tave.alarmissue.news.exceptions.VoteErrorCode.USER_ID_NOT_FOUND;
+
 
 @Service
 @RequiredArgsConstructor
@@ -34,11 +35,11 @@ public class NewsVoteService {
     public NewsVoteResponseDto createVoteAndGetResult(NewsVoteRequestDto dto, Long userId) {
         //유저가 없을때
         UserEntity user = userRepository.findById(userId)
-                .orElseThrow(() -> new VoteException(USER_ID_NOT_FOUND, "해당 유저를 찾을 수 없습니다."));
+                .orElseThrow(() -> new NewsVoteException(USER_ID_NOT_FOUND, "해당 유저를 찾을 수 없습니다."));
 
         //게시글이 없을때
         News news = newsRepository.findById(dto.getNewsId()).
-                orElseThrow(() -> new VoteException(POST_ID_NOT_FOUND, "postId: "+dto.getNewsId()));
+                orElseThrow(() -> new NewsVoteException(NEWS_ID_NOT_FOUND, "NewsId: "+ dto.getNewsId()));
 
         NewsVote vote=NewsVote.builder()
                 .news(news)
