+++ src/main/java/com/tave/alarmissue/post/controller/VoteController.java
@@ -1,0 +1,29 @@
+package com.tave.alarmissue.post.controller;
+
+import com.tave.alarmissue.auth.dto.request.PrincipalUserDetails;
+import com.tave.alarmissue.post.dto.request.VoteRequestDto;
+import com.tave.alarmissue.post.dto.response.VoteResponseDto;
+import com.tave.alarmissue.post.service.VoteService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/vote/v1")
+@RequiredArgsConstructor
+public class VoteController {
+
+    private final VoteService voteService;
+
+    @PostMapping //투표
+    public ResponseEntity<VoteResponseDto> createVoteAndGetResult(@RequestBody VoteRequestDto dto,
+                                                              @AuthenticationPrincipal PrincipalUserDetails principal) {
+        Long userId = principal.getUserId();
+        VoteResponseDto voteResponseDto = voteService.createVoteAndGetResult(dto, userId);
+
+        return ResponseEntity.status(HttpStatus.CREATED).body(voteResponseDto);
+    }
+
+}
