+++ src/main/java/com/tave/alarmissue/post/domain/Like.java
@@ -1,0 +1,49 @@
+package com.tave.alarmissue.post.domain;
+
+import com.tave.alarmissue.user.domain.UserEntity;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(name = "likes")
+public class Like {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long LikeId;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private UserEntity user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id", nullable = false)
+    private Post post;
+
+    @JoinColumn
+    private boolean liked;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "comment_id", nullable = true)
+    private Comment comment;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private LikeType likeType;
+
+    @Builder
+    public Like( UserEntity user, Post post,Comment comment, LikeType likeType,boolean liked) {
+        this.user = user;
+        this.post = post;
+        this.comment = comment;
+        this.likeType = likeType;
+        this.liked = liked;
+    }
+}
