+++ src/main/java/com/tave/alarmissue/post/domain/LikeType.java
@@ -1,0 +1,7 @@
+package com.tave.alarmissue.post.domain;
+
+public enum LikeType {
+    POST,
+    COMMENT
+    //REPLY
+}
