+++ src/main/java/com/tave/alarmissue/post/domain/VoteType.java
@@ -1,0 +1,7 @@
+package com.tave.alarmissue.post.domain;
+
+public enum VoteType {
+    BUY,    // 매수
+    HOLD,   // 보유
+    SELL    // 매도
+}
