+++ src/main/java/com/tave/alarmissue/post/dto/request/CommentCreateRequestDto.java
@@ -1,0 +1,10 @@
+package com.tave.alarmissue.post.dto.request;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class CommentCreateRequestDto {
+    private String commentContent;
+}
