+++ src/main/java/com/tave/alarmissue/post/dto/request/VoteRequestDto.java
@@ -1,0 +1,12 @@
+package com.tave.alarmissue.post.dto.request;
+
+import com.tave.alarmissue.post.domain.VoteType;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class VoteRequestDto {
+    private Long postId;
+    private VoteType voteType;
+}
