+++ src/main/java/com/tave/alarmissue/post/dto/response/CommentResponseDto.java
@@ -1,0 +1,20 @@
+package com.tave.alarmissue.post.dto.response;
+
+import com.tave.alarmissue.post.domain.VoteType;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@AllArgsConstructor
+@Builder
+public class CommentResponseDto {
+    private Long commentId;
+    private String commentContent;
+    private Long likeCount;
+    private String nickname;
+    private LocalDateTime createdAt;
+    private VoteType voteType;
+}
