+++ src/main/java/com/tave/alarmissue/post/dto/response/VoteCountResponse.java
@@ -1,0 +1,12 @@
+package com.tave.alarmissue.post.dto.response;
+
+import com.tave.alarmissue.post.domain.VoteType;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class VoteCountResponse {
+    private VoteType voteType;
+    private long count;
+}
