--- src/main/java/com/tave/alarmissue/vote/converter/VoteConverter.java
@@ -1,32 +1,0 @@
-package com.tave.alarmissue.vote.converter;
-
-import com.tave.alarmissue.post.domain.Post;
-import com.tave.alarmissue.vote.domain.VoteType;
-import com.tave.alarmissue.vote.dto.response.VoteCountResponse;
-import com.tave.alarmissue.vote.dto.response.VoteResponseDto;
-
-import java.util.List;
-
-public class VoteConverter {
-
-    public static VoteResponseDto toVoteResponseDto(Post post, VoteType myVoteType,
-                                                    List<VoteCountResponse> counts) {
-        int buyCount = 0, sellCount = 0, holdCount = 0;
-
-        for (VoteCountResponse c : counts) {
-            switch (c.getVoteType()) {
-                case BUY -> buyCount = (int) c.getCount();
-                case SELL -> sellCount = (int) c.getCount();
-                case HOLD -> holdCount = (int) c.getCount();
-            }
-        }
-
-        return VoteResponseDto.builder()
-                .postId(post.getPostId())
-                .buyCount(buyCount)
-                .sellCount(sellCount)
-                .holdCount(holdCount)
-                .voteType(myVoteType)
-                .build();
-    }
-}
