--- src/main/java/com/tave/alarmissue/vote/domain/VoteType.java
@@ -1,7 +1,0 @@
-package com.tave.alarmissue.vote.domain;
-
-public enum VoteType {
-    BUY,    // 매수
-    HOLD,   // 보유
-    SELL    // 매도
-}
