--- src/main/java/com/tave/alarmissue/vote/dto/request/VoteRequestDto.java
@@ -1,12 +1,0 @@
-package com.tave.alarmissue.vote.dto.request;
-
-import com.tave.alarmissue.vote.domain.VoteType;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class VoteRequestDto {
-    private Long postId;
-    private VoteType voteType;
-}
